# OpenAPI.yaml
openapi: 3.0.0

info:
  title: store
  version: '1.0'
  description: A simple store
  contact:
    name: SecuritEase Dev
    url: 'https://www.securitease.com'
    email: internal@securitease.com
  termsOfService: 'https://www.securitease.com'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.htm'

servers:
  - url: 'http://localhost:8080'
    description: dev

paths:
  /order:
    get:
      summary: Get Orders
      tags:
        - store
      operationId: getOrders
      description: Get all orders
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

    post:
      summary: Create Order
      tags:
        - store
      operationId: createOrder
      description: Create a new order using an existing customer and one or more existing products.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request, including a missing customer ID, blank description, or no products.
        '404':
          description: The specified customer or one or more products could not be found.

  /order/{id}:
    get:
      summary: Get Order By ID
      tags:
        - store
      operationId: getOrderById
      description: Get a specific order by its ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the order.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found.

  /customer:
    get:
      summary: Get Customers
      tags:
        - store
      operationId: getCustomers
      description: Get all customers, optionally filtering by a case-insensitive substring of the customer's name.
      parameters:
        - name: query
          in: query
          required: false
          description: Optional substring used to filter customer names. Matching is case-insensitive.
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'

    post:
      summary: Create Customer
      tags:
        - store
      operationId: createCustomer
      description: Create a new customer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                name:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

  /products:
    get:
      summary: Get Products
      tags:
        - store
      operationId: getProducts
      description: Get all products, including the IDs of orders containing each product.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      summary: Create Product
      tags:
        - store
      operationId: createProduct
      description: Create a new product. The product ID is generated by the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

  /products/{id}:
    get:
      summary: Get Product By ID
      tags:
        - store
      operationId: getProductById
      description: Get a specific product by its ID, including the IDs of orders containing it.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found.

components:
  schemas:
    Order:
      title: Order
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          $ref: '#/components/schemas/OrderCustomer'
        products:
          type: array
          items:
            $ref: '#/components/schemas/OrderProduct'

    OrderCustomer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string

    OrderProduct:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string

    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              description:
                type: string

    Product:
      title: Product
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        orderIds:
          type: array
          items:
            type: integer
            format: int64

    CreateOrderRequest:
      type: object
      required:
        - description
        - customerId
        - productIds
      properties:
        description:
          type: string
        customerId:
          type: integer
          format: int64
        productIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: integer
            format: int64

    CreateProductRequest:
      type: object
      required:
        - description
      properties:
        description:
          type: string

tags:
  - name: store
    description: store operations